Tezos Compiler!
This is a NodeJS program deployed in Amazon Web Services. It uses two S3 buckets and a Lambda. Any time a Tezos smart contract is added to the "in" S3 bucket, an event is triggered that wakes up the Lambda. The Lambda gets the smlse file from S3 and compiles the Tezos contract with it. This process creates 4 additional file (Storage, Type, Code and Code in JSON format). All of these files are then stored in another S3 bucket for use.